Reduce processing area for RasterCellIntersection
This commit improves RasterCellIntersection to more intelligently define
a processing region for a given polygon. Instead of considering the full
intersection of the raster extent and the polygon extent, it finds the
extent of the raster that is covered by some component of the polygon.
It also avoids processing polygon rings that do not intersect the
processing region.

In a rough performance test (total GPW population in Natural Earth 10m
countries) this improves performance by about 10%, from 300s to 260s.