Replace explicit GDALRasterWrapper with generic RasterSource
This allows for the implementation of other raster sources, e.g. one
dealing with R raster objects. Although returning an AbstractRaster from
the read_box method could be anticipated to cause a performance penalty
when the raster values are accessed, local testing does not show this to
be the case.