Fix incorrect result for polygon on grid boundary
When the boundary of a polygon exactly follows the raster grid, it may
be possible for a cell to be processed but have a covered fraction of
zero. This incorrectly allowed a flood fill to cross the border of the
raster and caused incorrect results.

The InitialSegmentOrientation work was an earlier attempt to resolve
this, but it was apparently not sufficiently robust. It may be possible
or desirable to remove that logic.

This bug was initially reported to the exactextractr R package at
https://github.com/isciences/exactextractr/issues/23